Configure user info in a .config block User info used to be set in a .run block. Moving it to a .config block prevents AuthService from prematurely capturing unconfigured user data

+++ viewer/annotations/comments.service.js
@@ -41,13 +41,15 @@
         }
 
         function deleteComment(comment) {
-            comment.delete().then(function success() {
-                var annotationComments = comments[comment.data.annotation_id];
-                var index = annotationComments.indexOf(comment);
-                annotationComments.splice(index, 1);
-            }, function error(response) {
-                console.log(response);
-            });
+            if (AuthService.deleteComment(comment)) {
+                comment.delete().then(function success() {
+                    var annotationComments = comments[comment.data.annotation_id];
+                    var index = annotationComments.indexOf(comment);
+                    annotationComments.splice(index, 1);
+                }, function error(response) {
+                    console.log(response);
+                });
+            }
         }
 
         return {
